requirements.txt
django
django-admin
django-cors-headers
requests
djangorestframework
youtube_transcript_api
transformers
# mysqlclient
gunicorn

